Use cell angles to find setting of hexagonal system (e.g. for "R 3")
Sometimes in the PDB, the setting of the hexagonal crystal system
is not clear from the symmetry symbol alone. For instance, the H-M symbol
"R 3" can mean either hexagonal or rhombohedral setting.
This ambiguity can be resolved by comparing angles of the unit cell.
The ratio of gamma to alpha angles is 120:90 in the hexagonal system
and 1:1 in rhombohedral. Therefore, find_spacegroup_by_name()
accepts now also alpha and gamma angles. If the angles are not passed,
the hexagonal system is returned.