faster reading of pdb files with millions of residues
It's for pseudo-PDB files from MD simulations - millions of residues.
Shouldn't notably affect reading of normal PDB files.

The slow part was finding a residue to which the atom should be
assigned. Because, just in case, the parser handles a corner case
when atoms from the same residue are not consecutive.
Perhaps this is an overkill and it should be assumed that atoms
from one residue are consecutive - I'm not sure.