mtz2cif: change _symmetry.space_group_name_H-M to pdb symbol
Rhombohedral groups in hexagonal settings start with H now.

The documentation for _symmetry.space_group_name_H-M says it is a
"symbol from International Tables for Crystallography Vol. A (2002)",
which would mean it can't be H 3, but on the other hand H 3
is listed there as allowed symbol and in commonly used in the archive.
This is probably because the PDBx/mmCIF spec was written partly
by the IUCr CIF commitee and partly by the PDB.