handle missing Entity::polymer_type
If _entity_poly.type is missing in mmCIF and neither ensure_entities()
nor setup_entities() is called - Entity::polymer_type is Unknown.
entity_type is used only in a couple places in the library,
so to be sure, check in these place if polymer type is known, and if not,
run check_polymer_type() to determine it.