cif2mtz: print warning if _refln has multirecord data (#261)
Warn if either Friedel pairs or multiple records for equivalent
reflections are found.
Anomalous data can be converted with --anomalous flag
(but for now this flag handles only F).

It can be either a mistake or an old way of depositing such data.
Nowadays different tags are used:
 _refln.pdbx_F_plus/minus etc for anomalous pairs,
_diffrn_refln for unmerged data.

Hopefully old-style data in the PDB will be eventually remediated.