add_hydrogens_without_positions(): fix occupancy in rare cases
and make the code clearer

When a residue with a partial occupancy (eg. A/200 FLR in 7euv)
has some atoms in two conformations, occupancy of riding Hs
on a parent that is bonded to an atom with two conformation,
but has only a single conformations (in such case we add Hs
in two conformations), may not need to be multiplied
by the occupancy of the parent.

Some other examples: 6ey1 3e9k 1e70 (but in these entries occupancies
might be erroneous).