map format: ignore ORIGIN (#288)
Previously, non-zero ORIGIN was causing the map to be regarded as not
fully set up. Not all properties were calculated, including the spacing.

Some functions may not work as expected if we just ignore the origin.
This might be improved in the future.  A note about it was added to docs.