build: fix gemmi-config.cmake and change defaults in pyproject.toml
gemmi-config.cmake didn't work if ZLIB was not found

and two unrelated changes:

in CMakeLists.txt, python module is marked as COMPONENT py,
so that it's possible to install only python module with:
cmake --install <dir> --component py

in pyproject.toml, setting BUILD_GEMMI_PROGRAM and INSTALL_DEV_FILES
to OFF was removed, so it's ON by default.
But it's set to off in cibuildwheel configuration.