update zlib library to ver. 1.3 to avoid warnings
This library is used only when zlib-ng is not used and
when zlib library is not installed in the system.

In the current version crc32.h is big: 580 KB.
I removed it and added -D DYNAMIC_CRC_TABLE.

Also, a fix to the previous commit: -I third_party/zlib was missing.