unitcell.hpp: separate parameters from UnitCell into UnitCellParameters
it's backward-compatible, UnitCell is derived from UnitCellParameters

UnitCell contains, in addition to the 6 parameters, two matrices
and a number of precalculated values. The new struct can be used
as a lightweight alternative - for passing parameters when other
values are not needed.

additionally, add a tiny helper function UnitCell::set_from_array()