read_pdb_from_stream: Allow wider MODEL field
Increases the maximum model sequence number to 99999999
(almost 100 million).

The format of a MODEL record in a PDB file is specified as the string
"MODEL ", some spaces and a 4-digit serial number. Due to the field
width of four, the highest model number can only be 9999.

However, this causes issues for some generated PDB files, such as converted MD
trajetories with more than 9999 frames.

This change "extends" the PDB format specificuation and uses the available
space currently occupied by spaces.