move spacegroup tables from header to src/symmetry.cpp
Comparing SpaceGroup pointers that should be pointing to the same entry
in gemmi's spacegroup table was failing. Apparently, gemmi shared library
and gemmi program had separate copies of the tables.

It was not a problem until some code that returns or assigns SpaceGroup
pointer was moved to the library.

The same thing is discussed here:
https://stackoverflow.com/questions/51374558/are-inline-variables-unique-across-boundaries