fix intermittant problem with find_tabulated_residue()
the static ResinfoData::array that's used was sometimes not initialized
on time (when python modules was linked with shared gemmi_cpp library).
and in the doctest float(130.0) is used instead of 130.0.
Not sure how it helps, it's not super clear why it helps.
With unknown=130.0 I sometimes get the result as if I passed unknown=0.